+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## Unreleased
+## 0.5.1
 
+- Mark ESLint v10 as supported
 - Support false positives with TypeScript function overloading (fixes [#105](https://github.com/ArnaudBarre/eslint-plugin-react-refresh/issues/105))
 - Support nested function calls for extraHOCs (fixes [#104](https://github.com/ArnaudBarre/eslint-plugin-react-refresh/issues/104))